+++ R/sys.R
@@ -1,5 +1,5 @@
 .tmax <- 100L
 .Lmax <- 1e+04
-.parameters <- c("tmax", "k", "Linf", "t0", "b", "Lm", "fb", "Rt",
+.parameters <- c("tmax", "k", "Linf", "t0", "Wb", "Lm", "fb", "Rt",
                 "nu", "Lv", "Vp", "Llo", "Lup", "Nc", "mu", "rho", "eta", "Rk",
                 "R0", "Wa", "fa")