+++ document.R
@@ -1,3 +1,2 @@
 devtools::document()
-codemetar::write_codemeta()
 devtools::check()