+++ NEWS.md
@@ -4,7 +4,8 @@
 - Added vulnerability power `Vp`.
 - Added slot limit non-complicance parameter `Nc`.
 - Increased upper limit for `Rk` from 10 to 100.
-- Added scaling parameter `R0`, and renamed `a` to `Wa`.
+- Added scaling parameter `R0`.
+- Renamed `a` to `Wa` and `b` to `Wb`.
 - Added `ypr_ggplot2()` to do customisable bivariate plot.
 
 # ypr 0.0.1